DL_Lab2_M11315051_林子新/requirements.txt
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
# nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pillow==11.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
# torch==2.4.0+cu124
# torchaudio==2.4.0+cu124
# torchvision==0.21.0+cu124
tqdm==4.67.1
#triton==3.2.0
typing_extensions==4.12.2
